pyproject.toml
[project]
name = "multisource-multimodal-rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.53.1",
]
